Extract relevant cases
