+++ src/test/java/reposense/parser/ArgsParserTest.java
@@ -21,13 +21,11 @@ import org.junit.Before;
 import org.junit.Test;
 
 import net.sourceforge.argparse4j.helper.HelpScreenException;
-import reposense.RepoSense;
 import reposense.model.CliArguments;
 import reposense.model.ConfigCliArguments;
 import reposense.model.FileType;
 import reposense.model.FileTypeTest;
 import reposense.model.LocationsCliArguments;
-import reposense.model.RepoConfiguration;
 import reposense.model.ViewCliArguments;
 import reposense.util.FileUtil;
 import reposense.util.InputBuilder;
@@ -52,8 +50,7 @@ public class ArgsParserTest {
     private static final InputBuilder DEFAULT_INPUT_BUILDER = new InputBuilder();
 
     private static final String TEST_REPO_REPOSENSE = "https://github.com/reposense/RepoSense.git";
-    private static final String TEST_REPO_BETA = "https://github.com/reposense/testrepo-Beta.git";
-    private static final String TEST_REPO_CHARLIE = "https://github.com/reposense/testrepo-Charlie.git";
+    private static final String TEST_REPO_BETA = "https://github.com/reposense/testrepo-Beta.git";;
     private static final String TEST_REPO_DELTA = "https://github.com/reposense/testrepo-Delta.git";
 
     private static final String DEFAULT_TIMEZONE = "Asia/Singapore";
@@ -452,15 +449,6 @@ public class ArgsParserTest {
     }
 
     @Test
-    public void parse_validGitRepoLocations_repoConfigurationListCorrectSize() throws Exception {
-        String input = new InputBuilder().addRepos(TEST_REPO_REPOSENSE, TEST_REPO_DELTA).build();
-        CliArguments cliArguments = ArgsParser.parse(translateCommandline(input));
-        Assert.assertTrue(cliArguments instanceof LocationsCliArguments);
-        List<RepoConfiguration> repoConfigs = RepoSense.getRepoConfigurations((LocationsCliArguments) cliArguments);
-        Assert.assertEquals(2, repoConfigs.size());
-    }
-
-    @Test
     public void parse_repoLocationsAndView_returnsLocationCliArguments() throws Exception {
         String input = new InputBuilder().addRepos(TEST_REPO_REPOSENSE, TEST_REPO_DELTA)
                 .addView()
@@ -506,23 +494,6 @@ public class ArgsParserTest {
     }
 
     @Test
-    public void parse_configOrLocationsSimilar_success() throws Exception {
-        String input = new InputBuilder().addConfig(CONFIG_FOLDER_ABSOLUTE).build();
-        CliArguments configCliArguments = ArgsParser.parse(translateCommandline(input));
-        Assert.assertTrue(configCliArguments instanceof ConfigCliArguments);
-        List<RepoConfiguration> actualRepoConfigs =
-                RepoSense.getRepoConfigurations((ConfigCliArguments) configCliArguments);
-
-        input = new InputBuilder().addRepos(TEST_REPO_BETA, TEST_REPO_CHARLIE, TEST_REPO_DELTA).build();
-        CliArguments locationCliArguments = ArgsParser.parse(translateCommandline(input));
-        Assert.assertTrue(locationCliArguments instanceof LocationsCliArguments);
-        List<RepoConfiguration> expectedRepoConfigs =
-                RepoSense.getRepoConfigurations((LocationsCliArguments) locationCliArguments);
-
-        Assert.assertEquals(actualRepoConfigs, expectedRepoConfigs);
-    }
-
-    @Test
     public void emptyArgs_defaultConfigFolderPath() throws Exception {
         CliArguments cliArguments = ArgsParser.parse(new String[]{});
         Assert.assertTrue(cliArguments instanceof ConfigCliArguments);
@@ -539,15 +510,6 @@ public class ArgsParserTest {
         CliArguments reposAliasCliArguments = ArgsParser.parse(translateCommandline(input));
 
         Assert.assertEquals(repoAliasCliArguments, reposAliasCliArguments);
-    }
-
-    @Test (expected = ParseException.class)
-    public void parse_noValidRepoLocation_throwsParseException()
-            throws ParseException, HelpScreenException {
-        String input = new InputBuilder().addRepos("https://githubaaaa.com/asdasdasdasd/RepoSense").build();
-        CliArguments cliArguments = ArgsParser.parse(translateCommandline(input));
-        Assert.assertTrue(cliArguments instanceof LocationsCliArguments);
-        RepoSense.getRepoConfigurations((LocationsCliArguments) cliArguments);
     }
 
     @Test(expected = ParseException.class)
